+++ pyengine/Network/__init__.py
@@ -1,2 +1,1 @@
 from pyengine.Network.NetworkManager import NetworkManager
-from pyengine.Network.Packet import Packet